+++ src/main/java/com/loha/flippedclassroom/dao/SeminarDao.java
@@ -38,7 +38,7 @@ public class SeminarDao {
     /**
      * 根据Id获取当前的讨论课
      */
-    public Seminar getCurSeminar(Integer seminarId) throws Exception{
+    public Seminar getCurSeminar(Long seminarId) throws Exception{
         return seminarMapper.selectSeminarById(seminarId);
     }
 }
