+++ src/main/java/com/loha/flippedclassroom/entity/Klass.java
@@ -12,10 +12,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Klass {
-    private Integer id;
-    private Integer courseId;
+    private Long id;
+    private Long courseId;
     private Integer grade;
     private Integer klassSerial;
     private String time;
     private String location;
+
 }
