+++ src/main/java/com/loha/flippedclassroom/mapper/TeamMapper.java
@@ -20,7 +20,7 @@ public interface TeamMapper {
      * @return a team
      * @throws Exception
      */
-    Team selectTeamById(Integer teamId) throws Exception;
+    Team selectTeamById(Long teamId) throws Exception;
 
     /**
      * 查找一个课程下所有的team
@@ -28,5 +28,5 @@ public interface TeamMapper {
      * @return team list
      * @throws Exception
      */
-    List<Team> selectTeamByCourseId(Integer courseId) throws Exception;
+    List<Team> selectTeamByCourseId(Long courseId) throws Exception;
 }
